backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic
python-dotenv
